Remove plateGrid from SampleChanger redux chunk
plateGrid introduced only set of constants. There is no functionality in the current
code base that modifies it. All related code  moved to seperate file.
